@@ -26,26 +26,44 @@
     "clean": "rimraf lib",
     "clean:examples": "rimraf examples/static",
     "l": "lessc --autoprefix=\"ie >= 10, safari >= 8, last 2 versions\" ",
-    "less": "npm run l src/less/styles.less ./lib/css/react-big-calendar.css && npm run less-dnd",
-    "less-dnd": "npm run l src/addons/dragAndDrop/styles.less ./lib/addons/dragAndDrop/styles.css",
+    "less":
+      "npm run l src/less/styles.less ./lib/css/react-big-calendar.css && npm run less-dnd",
+    "less-dnd":
+      "npm run l src/addons/dragAndDrop/styles.less ./lib/addons/dragAndDrop/styles.css",
     "assets": "cpy src/less/* lib/less && npm run assets-addons",
     "assets-addons": "cpy addons/**/*.less ../lib/ --cwd=src --parents",
-    "build": "npm run clean && babel src --out-dir lib && npm run assets && npm run less && npm run less",
-    "build:examples": "npm run clean:examples && webpack --config webpack/examples.config.js",
-    "build:visual-test": "webpack --config webpack/visual-test.js",
-    "examples": "npm run clean:examples && webpack-dev-server --inline --hot --config webpack/examples.config.js",
-    "lint": "eslint src --ext .jsx --ext .js",
+    "build":
+      "npm run clean && babel src --out-dir lib && npm run assets && npm run less && npm run less",
+    "build:examples":
+      "npm run clean:examples && webpack --config examples/webpack.config.js",
+    "examples":
+      "npm run clean:examples && webpack-dev-server --inline --hot --config examples/webpack.config.js",
+    "lint": "eslint src test",
     "storybook": "start-storybook -p 9002",
-    "test": "npm run lint",
-    "tdd": "karma start",
-    "release": "release"
+    "test": "npm run lint && jest",
+    "tdd": "jest --watch",
+    "release": "release",
+    "prettier": "prettier '**/*js' !examples/bundle.js '!lib/**' --write ",
+    "precommit": "lint-staged"
+  },
+  "lint-staged": {
+    "src/**/*.js": "eslint",
+    "test/**/*.js": "eslint",
+    "*.{js,json,css,md}": ["prettier --write", "git add"]
+  },
+  "prettier": {
+    "printWidth": 80,
+    "semi": false,
+    "singleQuote": true,
+    "trailingComma": "es5",
+    "bracketSpacing": true
   },
   "peerDependencies": {
     "react": "^15.3.0 || ^16.0.0",
     "react-dom": "^15.3.0 || ^16.0.0"
   },
   "devDependencies": {
-    "@kadira/storybook": "^1.28.1",
+    "@storybook/react": "3.2.16",
     "babel-cli": "^6.7.5",
     "babel-core": "^6.7.6",
     "babel-eslint": "^6.0.2",
@@ -55,27 +73,33 @@
     "cpy": "^3.4.1",
     "eslint": "^3.0.0",
     "eslint-config-jason": "^4.0.0",
+    "eslint-config-prettier": "^2.9.0",
     "eslint-import-resolver-webpack": "^0.8.0",
     "eslint-plugin-import": "^2.0.0",
     "eslint-plugin-react": "^6.8.0",
     "globalize": "^0.1.1",
-    "less": "^2.7.2",
+    "husky": "^0.14.3",
+    "jest": "^21.1.0",
+    "less": "^2.7.3",
+    "less-loader": "^4.0.5",
     "less-plugin-autoprefix": "^1.5.1",
-    "markdown-jsx-loader": "^2.0.1",
+    "lint-staged": "^6.0.0",
+    "markdown-jsx-loader": "^3.0.2",
     "marked": "^0.3.5",
     "moment": "^2.17.1",
     "mt-changelog": "^0.6.1",
-    "react": "^15.3.0",
+    "prettier": "^1.9.1",
+    "react": "^16.2.0",
     "react-bootstrap": "^0.28.0",
     "react-dnd": "^2.1.4",
     "react-dnd-html5-backend": "^2.1.2",
     "react-docgen": "^2.14.1",
-    "react-dom": "^15.3.0",
+    "react-dom": "^16.2.0",
     "release-script": "^1.0.0",
     "rimraf": "^2.4.2",
-    "webpack": "^2.4.1",
-    "webpack-atoms": "^1.0.2",
-    "webpack-dev-server": "^2.4.5"
+    "webpack": "^3.10.0",
+    "webpack-atoms": "^4.1.2",
+    "webpack-dev-server": "^2.9.5"
   },
   "dependencies": {
     "classnames": "^2.1.3",
@@ -85,7 +109,7 @@
     "prop-types": "^15.5.8",
     "react-overlays": "^0.8.3",
     "react-prop-types": "^0.4.0",
-    "uncontrollable": "^3.3.1 || ^4.0.0",
+    "uncontrollable": "^4.0.0",
     "warning": "^2.0.0"
   }
 }
