@@ -1,38 +1,19 @@
-const path = require('path');
-const Autoprefixer = require('less-plugin-autoprefix');
+const { rules, loaders, plugins, stats } = require('webpack-atoms')
 
-module.exports = {
-  resolve: {
-    extensions: ['', '.js', '.jsx']
-  },
-  module: {
-    loaders: [
-      {
-        test: /\.css?$/,
-        loaders: [ 'style', 'raw' ],
-        include: [
-          path.resolve(__dirname, '../')
-        ]
-      },
-      {
-        test: /\.less?$/,
-        loader: 'style-loader!css-loader!less-loader',
-        include: [
-          path.resolve(__dirname, '../')
-        ]
-      },
-      {test: /\.(woff|woff2)(\?v=\d+\.\d+\.\d+)?$/, loader: 'url?limit=10000&mimetype=application/font-woff'},
-      {test: /\.ttf(\?v=\d+\.\d+\.\d+)?$/, loader: 'url?limit=10000&mimetype=application/octet-stream'},
-      {test: /\.eot(\?v=\d+\.\d+\.\d+)?$/, loader: 'file'},
-      {test: /\.svg(\?v=\d+\.\d+\.\d+)?$/, loader: 'url?limit=10000&mimetype=image/svg+xml'}
-    ]
-  },
+const browsers = ['last 2 versions', 'ie >= 10']
 
-  lessLoader: {
-    lessPlugins: [
-      new Autoprefixer({
-        browsers: ['last 2 versions', 'ie >= 10']
-      })
-    ]
+module.exports = function(config) {
+  return {
+    ...config,
+    module: {
+      rules: [
+        rules.js(),
+        rules.fonts(),
+        rules.images(),
+        rules.css(),
+        rules.less({ browsers }),
+      ],
+    },
+    plugins: [...config.plugins, plugins.extractText()],
   }
-};
+}
