@@ -1,22 +1,21 @@
-import React from 'react';
+import React from 'react'
 
-import dates from 'date-arithmetic';
-import events from '../events';
-import BigCalendar from 'react-big-calendar';
-import localizer from 'react-big-calendar/lib/localizer';
-import TimeGrid from 'react-big-calendar/lib/TimeGrid';
-
+import dates from 'date-arithmetic'
+import events from '../events'
+import BigCalendar from 'react-big-calendar'
+import localizer from 'react-big-calendar/lib/localizer'
+import TimeGrid from 'react-big-calendar/lib/TimeGrid'
 
 const getRange = (date, culture) => {
-  let firstOfWeek = localizer.startOfWeek(culture);
-  let start = dates.startOf(date, 'week', firstOfWeek);
-  let end = dates.endOf(date, 'week', firstOfWeek);
+  let firstOfWeek = localizer.startOfWeek(culture)
+  let start = dates.startOf(date, 'week', firstOfWeek)
+  let end = dates.endOf(date, 'week', firstOfWeek)
 
   if (firstOfWeek === 1) {
-    end = dates.subtract(end, 2, 'day');
+    end = dates.subtract(end, 2, 'day')
   } else {
-    start = dates.add(start, 1, 'day');
-    end = dates.subtract(end, 1, 'day');
+    start = dates.add(start, 1, 'day')
+    end = dates.subtract(end, 1, 'day')
   }
 
   return dates.range(start, end)
@@ -24,25 +23,23 @@ const getRange = (date, culture) => {
 
 class MyWeek extends React.Component {
   render() {
-    let { date, culture } = this.props;
-    let range = getRange(date, culture);
+    let { date, culture } = this.props
+    let range = getRange(date, culture)
 
-    return (
-      <TimeGrid {...this.props} range={range} eventOffset={15} />
-    );
+    return <TimeGrid {...this.props} range={range} eventOffset={15} />
   }
 }
 
 MyWeek.navigate = (date, action) => {
-  switch (action){
+  switch (action) {
     case BigCalendar.Navigate.PREVIOUS:
-      return dates.add(date, -1, 'week');
+      return dates.add(date, -1, 'week')
 
     case BigCalendar.Navigate.NEXT:
       return dates.add(date, 1, 'week')
 
     default:
-      return date;
+      return date
   }
 }
 
@@ -50,20 +47,15 @@ MyWeek.title = (date, { formats, culture }) => {
   return `My awesome week: ${Date.toLocaleString()}`
 }
 
+let CustomView = props => (
+  <div>
+    <BigCalendar
+      events={events}
+      defaultDate={new Date(2015, 3, 1)}
+      views={{ month: true, week: MyWeek }}
+      test="io"
+    />
+  </div>
+)
 
-let CustomView = React.createClass({
-  render(){
-    return (
-      <div>
-        <BigCalendar
-          events={events}
-          defaultDate={new Date(2015, 3, 1)}
-          views={{ month: true, week: MyWeek }}
-          test="io"
-        />
-      </div>
-    )
-  }
-})
-
-export default CustomView;
+export default CustomView
