@@ -3,45 +3,61 @@ import events from '../events'
 import HTML5Backend from 'react-dnd-html5-backend'
 import { DragDropContext } from 'react-dnd'
 import BigCalendar from 'react-big-calendar'
-import withDragAndDrop from 'react-big-calendar/lib/addons/dragAndDrop';
+import withDragAndDrop from 'react-big-calendar/lib/addons/dragAndDrop'
 
-import 'react-big-calendar/lib/addons/dragAndDrop/styles.less';
+import 'react-big-calendar/lib/addons/dragAndDrop/styles.less'
 
-const DragAndDropCalendar = withDragAndDrop(BigCalendar);
+const DragAndDropCalendar = withDragAndDrop(BigCalendar)
 
 class Dnd extends React.Component {
-  constructor (props) {
+  constructor(props) {
     super(props)
     this.state = {
-      events: events
+      events: events,
     }
 
     this.moveEvent = this.moveEvent.bind(this)
   }
 
   moveEvent({ event, start, end }) {
-    const { events } = this.state;
+    const { events } = this.state
 
-    const idx = events.indexOf(event);
-    const updatedEvent = { ...event, start, end };
+    const idx = events.indexOf(event)
+    const updatedEvent = { ...event, start, end }
 
     const nextEvents = [...events]
     nextEvents.splice(idx, 1, updatedEvent)
 
     this.setState({
-      events: nextEvents
+      events: nextEvents,
     })
 
-    alert(`${event.title} was dropped onto ${event.start}`);
+    alert(`${event.title} was dropped onto ${event.start}`)
   }
 
-  render(){
+  resizeEvent = (resizeType, { event, start, end }) => {
+    const { events } = this.state
+
+    const nextEvents = events.map(existingEvent => {
+      return existingEvent.id == event.id
+        ? { ...existingEvent, start, end }
+        : existingEvent
+    })
+
+    this.setState({
+      events: nextEvents,
+    })
+  }
+
+  render() {
     return (
       <DragAndDropCalendar
         selectable
         events={this.state.events}
         onEventDrop={this.moveEvent}
-        defaultView='week'
+        resizable
+        onEventResize={this.resizeEvent}
+        defaultView="week"
         defaultDate={new Date(2015, 3, 12)}
       />
     )
