@@ -1,20 +1,16 @@
-import React from 'react';
-import BigCalendar from 'react-big-calendar';
-import events from '../events';
+import React from 'react'
+import BigCalendar from 'react-big-calendar'
+import events from '../events'
 
-let Timeslots = React.createClass({
-  render(){
-    return (
-      <BigCalendar
-        {...this.props}
-        events={events}
-        step={15}
-        timeslots={8}
-        defaultView='week'
-        defaultDate={new Date(2015, 3, 12)}
-      />
-    )
-  }
-})
+let Timeslots = props => (
+  <BigCalendar
+    {...props}
+    events={events}
+    step={15}
+    timeslots={8}
+    defaultView="week"
+    defaultDate={new Date(2015, 3, 12)}
+  />
+)
 
-export default Timeslots;
+export default Timeslots
