@@ -1,6 +1,6 @@
-const webpack = require('webpack');
-const WebpackDevServer = require('webpack-dev-server');
-const config = require('../webpack/examples.config.es6');
+const webpack = require('webpack')
+const WebpackDevServer = require('webpack-dev-server')
+const config = require('../webpack/examples.config.es6')
 
 new WebpackDevServer(webpack(config), {
   publicPath: config.output.publicPath,
@@ -9,12 +9,12 @@ new WebpackDevServer(webpack(config), {
   quiet: false,
   progress: true,
   stats: {
-    colors: true
-  }
-}).listen(config.port, 'localhost', function (err) {
+    colors: true,
+  },
+}).listen(config.port, 'localhost', function(err) {
   if (err) {
-    console.log(err);
+    console.log(err)
   } else {
-    console.log('Listening at localhost:' + config.port);
+    console.log('Listening at localhost:' + config.port)
   }
-});
+})
