@@ -1,18 +1,23 @@
-import cn from 'classnames';
-import getHeight from 'dom-helpers/query/height';
-import qsa from 'dom-helpers/query/querySelectorAll';
-import PropTypes from 'prop-types';
-import React from 'react';
-import { findDOMNode } from 'react-dom';
+import cn from 'classnames'
+import getHeight from 'dom-helpers/query/height'
+import qsa from 'dom-helpers/query/querySelectorAll'
+import PropTypes from 'prop-types'
+import React from 'react'
+import { findDOMNode } from 'react-dom'
 
-import dates from './utils/dates';
-import { accessor, elementType } from './utils/propTypes';
-import { segStyle, eventSegments, endOfRange, eventLevels } from './utils/eventLevels';
-import BackgroundCells from './BackgroundCells';
-import EventRow from './EventRow';
-import EventEndingRow from './EventEndingRow';
+import dates from './utils/dates'
+import { accessor, elementType } from './utils/propTypes'
+import {
+  segStyle,
+  eventSegments,
+  endOfRange,
+  eventLevels,
+} from './utils/eventLevels'
+import BackgroundCells from './BackgroundCells'
+import EventRow from './EventRow'
+import EventEndingRow from './EventEndingRow'
 
-let isSegmentInSlot = (seg, slot) => seg.left <= slot && seg.right >= slot;
+let isSegmentInSlot = (seg, slot) => seg.left <= slot && seg.right >= slot
 
 const propTypes = {
   date: PropTypes.instanceOf(Date),
@@ -32,6 +37,7 @@ const propTypes = {
   onSelectSlot: PropTypes.func,
   onSelectEnd: PropTypes.func,
   onSelectStart: PropTypes.func,
+  dayPropGetter: PropTypes.func,
 
   now: PropTypes.instanceOf(Date).isRequired,
   startAccessor: accessor.isRequired,
@@ -42,7 +48,7 @@ const propTypes = {
   eventWrapperComponent: elementType.isRequired,
   minRows: PropTypes.number.isRequired,
   maxRows: PropTypes.number.isRequired,
-};
+}
 
 const defaultProps = {
   minRows: 0,
@@ -50,57 +56,53 @@ const defaultProps = {
 }
 
 class DateContentRow extends React.Component {
-
   constructor(...args) {
-    super(...args);
+    super(...args)
   }
 
-  handleSelectSlot = (slot) => {
-    const { range, onSelectSlot } = this.props;
+  handleSelectSlot = slot => {
+    const { range, onSelectSlot } = this.props
 
-    onSelectSlot(
-      range.slice(slot.start, slot.end + 1),
-      slot,
-    )
+    onSelectSlot(range.slice(slot.start, slot.end + 1), slot)
   }
 
-  handleShowMore = (slot) => {
-    const { range, onShowMore } = this.props;
+  handleShowMore = slot => {
+    const { range, onShowMore } = this.props
     let row = qsa(findDOMNode(this), '.rbc-row-bg')[0]
 
-    let cell;
-    if (row) cell = row.children[slot-1]
+    let cell
+    if (row) cell = row.children[slot - 1]
 
     let events = this.segments
       .filter(seg => isSegmentInSlot(seg, slot))
       .map(seg => seg.event)
 
-    onShowMore(events, range[slot-1], cell, slot)
+    onShowMore(events, range[slot - 1], cell, slot)
   }
 
   createHeadingRef = r => {
-    this.headingRow = r;
+    this.headingRow = r
   }
 
   createEventRef = r => {
-    this.eventRow = r;
+    this.eventRow = r
   }
 
   getContainer = () => {
-    const { container } = this.props;
+    const { container } = this.props
     return container ? container() : findDOMNode(this)
   }
 
   getRowLimit() {
-    let eventHeight = getHeight(this.eventRow);
+    let eventHeight = getHeight(this.eventRow)
     let headingHeight = this.headingRow ? getHeight(this.headingRow) : 0
-    let eventSpace = getHeight(findDOMNode(this)) - headingHeight;
+    let eventSpace = getHeight(findDOMNode(this)) - headingHeight
 
     return Math.max(Math.floor(eventSpace / eventHeight), 1)
   }
 
   renderHeadingCell = (date, index) => {
-    let { renderHeader, range } = this.props;
+    let { renderHeader, range } = this.props
 
     return renderHeader({
       date,
@@ -108,25 +110,25 @@ class DateContentRow extends React.Component {
       style: segStyle(1, range.length),
       className: cn(
         'rbc-date-cell',
-        dates.eq(date, this.props.now, 'day') && 'rbc-now', // FIXME use props.now
-      )
+        dates.eq(date, this.props.now, 'day') && 'rbc-now' // FIXME use props.now
+      ),
     })
   }
 
   renderDummy = () => {
-    let { className, range, renderHeader } = this.props;
+    let { className, range, renderHeader } = this.props
     return (
       <div className={className}>
-        <div className='rbc-row-content'>
+        <div className="rbc-row-content">
           {renderHeader && (
-            <div className='rbc-row' ref={this.createHeadingRef}>
+            <div className="rbc-row" ref={this.createHeadingRef}>
               {range.map(this.renderHeadingCell)}
             </div>
           )}
-          <div className='rbc-row' ref={this.createEventRef}>
-            <div className='rbc-row-segment' style={segStyle(1, range.length)}>
-              <div className='rbc-event'>
-                <div className='rbc-event-content'>&nbsp;</div>
+          <div className="rbc-row" ref={this.createEventRef}>
+            <div className="rbc-row-segment" style={segStyle(1, range.length)}>
+              <div className="rbc-event">
+                <div className="rbc-event-content">&nbsp;</div>
               </div>
             </div>
           </div>
@@ -143,11 +145,13 @@ class DateContentRow extends React.Component {
       range,
       className,
       selectable,
+      dayPropGetter,
       renderForMeasure,
       startAccessor,
       endAccessor,
       renderHeader,
-      minRows, maxRows,
+      minRows,
+      maxRows,
       dateCellWrapper,
       eventComponent,
       eventWrapperComponent,
@@ -155,20 +159,27 @@ class DateContentRow extends React.Component {
       onSelectEnd,
       longPressThreshold,
       ...props
-    } = this.props;
+    } = this.props
 
-    if (renderForMeasure)
-      return this.renderDummy();
+    if (renderForMeasure) return this.renderDummy()
 
-    let { first, last } = endOfRange(range);
+    let { first, last } = endOfRange(range)
 
-    let segments = this.segments = events.map(evt => eventSegments(evt, first, last, {
-      startAccessor,
-      endAccessor
-    }, range))
+    let segments = (this.segments = events.map(evt =>
+      eventSegments(
+        evt,
+        first,
+        last,
+        {
+          startAccessor,
+          endAccessor,
+        },
+        range
+      )
+    ))
 
-    let { levels, extra } = eventLevels(segments, Math.max(maxRows - 1, 1));
-    while (levels.length < minRows ) levels.push([])
+    let { levels, extra } = eventLevels(segments, Math.max(maxRows - 1, 1))
+    while (levels.length < minRows) levels.push([])
 
     return (
       <div className={className}>
@@ -178,6 +189,7 @@ class DateContentRow extends React.Component {
           range={range}
           selectable={selectable}
           container={this.getContainer}
+          dayPropGetter={dayPropGetter}
           onSelectStart={onSelectStart}
           onSelectEnd={onSelectEnd}
           onSelectSlot={this.handleSelectSlot}
@@ -185,13 +197,13 @@ class DateContentRow extends React.Component {
           longPressThreshold={longPressThreshold}
         />
 
-        <div className='rbc-row-content'>
+        <div className="rbc-row-content">
           {renderHeader && (
-            <div className='rbc-row' ref={this.createHeadingRef}>
+            <div className="rbc-row" ref={this.createHeadingRef}>
               {range.map(this.renderHeadingCell)}
             </div>
           )}
-          {levels.map((segs, idx) =>
+          {levels.map((segs, idx) => (
             <EventRow
               {...props}
               key={idx}
@@ -204,7 +216,7 @@ class DateContentRow extends React.Component {
               startAccessor={startAccessor}
               endAccessor={endAccessor}
             />
-          )}
+          ))}
           {!!extra.length && (
             <EventEndingRow
               {...props}
@@ -220,11 +232,11 @@ class DateContentRow extends React.Component {
           )}
         </div>
       </div>
-    );
+    )
   }
 }
 
-DateContentRow.propTypes = propTypes;
-DateContentRow.defaultProps = defaultProps;
+DateContentRow.propTypes = propTypes
+DateContentRow.defaultProps = defaultProps
 
 export default DateContentRow
