@@ -1,16 +1,16 @@
-import PropTypes from 'prop-types';
+import PropTypes from 'prop-types'
 import React from 'react'
 
 const DateHeader = ({ label, drilldownView, onDrillDown }) => {
-    if (!drilldownView) {
-      return (<span>{label}</span>)
-    }
+  if (!drilldownView) {
+    return <span>{label}</span>
+  }
 
-    return (
-      <a href="#" onClick={onDrillDown}>
-        {label}
-      </a>
-    )
+  return (
+    <a href="#" onClick={onDrillDown}>
+      {label}
+    </a>
+  )
 }
 
 DateHeader.propTypes = {
@@ -18,7 +18,7 @@ DateHeader.propTypes = {
   date: PropTypes.instanceOf(Date),
   drilldownView: PropTypes.string,
   onDrillDown: PropTypes.func,
-  isOffRange: PropTypes.bool
+  isOffRange: PropTypes.bool,
 }
 
 export default DateHeader
