@@ -1,41 +1,43 @@
-import PropTypes from 'prop-types';
-import React from 'react';
+import PropTypes from 'prop-types'
+import React from 'react'
 
-import dates from './utils/dates';
-import { navigate } from './utils/constants';
-import TimeGrid from './TimeGrid';
-import localizer  from './localizer';
+import dates from './utils/dates'
+import { navigate } from './utils/constants'
+import TimeGrid from './TimeGrid'
+import localizer from './localizer'
 
 class Day extends React.Component {
   static propTypes = {
     date: PropTypes.instanceOf(Date).isRequired,
-  };
+  }
 
   render() {
-    let { date, ...rest } = this.props;
+    const { date, ...rest } = this.props
 
     return (
-      <TimeGrid {...rest} range={[dates.startOf(date, 'day')]} eventOffset={10} />
-    );
+      <TimeGrid
+        {...rest}
+        range={[dates.startOf(date, 'day')]}
+        eventOffset={10}
+      />
+    )
   }
 }
 
-Day.navigate = (date, action)=>{
-  switch (action){
+Day.navigate = (date, action) => {
+  switch (action) {
     case navigate.PREVIOUS:
-      return dates.add(date, -1, 'day');
+      return dates.add(date, -1, 'day')
 
     case navigate.NEXT:
       return dates.add(date, 1, 'day')
 
     default:
-      return date;
+      return date
   }
 }
 
-
 Day.title = (date, { formats, culture }) =>
-  localizer.format(date, formats.dayHeaderFormat, culture);
-
+  localizer.format(date, formats.dayHeaderFormat, culture)
 
 export default Day
