@@ -1,23 +1,26 @@
-import PropTypes from 'prop-types';
-import React from 'react';
-import { findDOMNode } from 'react-dom';
-import cn from 'classnames';
+import PropTypes from 'prop-types'
+import React from 'react'
+import { findDOMNode } from 'react-dom'
+import cn from 'classnames'
 
-import Selection, { getBoundsForNode, isEvent } from './Selection';
-import dates from './utils/dates';
-import { isSelected } from './utils/selection';
+import Selection, { getBoundsForNode, isEvent } from './Selection'
+import dates from './utils/dates'
+import { isSelected } from './utils/selection'
 import localizer from './localizer'
 
-import { notify } from './utils/helpers';
-import { accessor, elementType, dateFormat } from './utils/propTypes';
-import { accessor as get } from './utils/accessors';
+import { notify } from './utils/helpers'
+import { accessor, elementType, dateFormat } from './utils/propTypes'
+import { accessor as get } from './utils/accessors'
 
-import getStyledEvents, { positionFromDate, startsBefore } from './utils/dayViewLayout'
+import getStyledEvents, {
+  positionFromDate,
+  startsBefore,
+} from './utils/dayViewLayout'
 
 import TimeColumn from './TimeColumn'
 
-function snapToSlot(date, step){
-  var roundTo = 1000 * 60 * step;
+function snapToSlot(date, step) {
+  var roundTo = 1000 * 60 * step
   return new Date(Math.floor(date.getTime() / roundTo) * roundTo)
 }
 
@@ -28,6 +31,7 @@ function startsAfter(date, max) {
 class DayColumn extends React.Component {
   static propTypes = {
     events: PropTypes.array.isRequired,
+    components: PropTypes.object,
     step: PropTypes.number.isRequired,
     min: PropTypes.instanceOf(Date).isRequired,
     max: PropTypes.instanceOf(Date).isRequired,
@@ -61,32 +65,32 @@ class DayColumn extends React.Component {
     className: PropTypes.string,
     dragThroughEvents: PropTypes.bool,
     eventPropGetter: PropTypes.func,
+    dayPropGetter: PropTypes.func,
     dayWrapperComponent: elementType,
     eventComponent: elementType,
     eventWrapperComponent: elementType.isRequired,
-  };
+    resource: PropTypes.string,
+  }
 
   static defaultProps = {
     dragThroughEvents: true,
     timeslots: 2,
-  };
+  }
 
-  state = { selecting: false };
+  state = { selecting: false }
 
   componentDidMount() {
-    this.props.selectable
-    && this._selectable()
+    this.props.selectable && this._selectable()
   }
 
   componentWillUnmount() {
-    this._teardownSelectable();
+    this._teardownSelectable()
   }
 
   componentWillReceiveProps(nextProps) {
-    if (nextProps.selectable && !this.props.selectable)
-      this._selectable();
+    if (nextProps.selectable && !this.props.selectable) this._selectable()
     if (!nextProps.selectable && this.props.selectable)
-      this._teardownSelectable();
+      this._teardownSelectable()
   }
 
   render() {
@@ -97,103 +101,122 @@ class DayColumn extends React.Component {
       now,
       selectRangeFormat,
       culture,
+      dayPropGetter,
       ...props
     } = this.props
 
     this._totalMin = dates.diff(min, max, 'minutes')
     let { selecting, startSlot, endSlot } = this.state
-    let style = this._slotStyle(startSlot, endSlot)
+    let slotStyle = this._slotStyle(startSlot, endSlot)
 
     let selectDates = {
       start: this.state.startDate,
-      end: this.state.endDate
-    };
+      end: this.state.endDate,
+    }
+
+    const { className, style } = (dayPropGetter && dayPropGetter(max)) || {}
 
     return (
       <TimeColumn
         {...props}
         className={cn(
           'rbc-day-slot',
+          className,
           dates.isToday(max) && 'rbc-today'
         )}
+        style={style}
         now={now}
         min={min}
         max={max}
         step={step}
       >
-        {this.renderEvents()}
-
-        {selecting &&
-          <div className='rbc-slot-selection' style={style}>
-              <span>
-              { localizer.format(selectDates, selectRangeFormat, culture) }
-              </span>
+        <div className={cn('rbc-events-container', { rtl: this.props.rtl })}>
+          {this.renderEvents()}
+        </div>
+        {selecting && (
+          <div className="rbc-slot-selection" style={slotStyle}>
+            <span>
+              {localizer.format(selectDates, selectRangeFormat, culture)}
+            </span>
           </div>
-        }
+        )}
       </TimeColumn>
-    );
+    )
   }
 
   renderEvents = () => {
     let {
-        events
-      , min
-      , max
-      , showMultiDayTimes
-      , culture
-      , eventPropGetter
-      , selected
-      , messages
-      , eventComponent
-      , eventTimeRangeFormat
-      , eventTimeRangeStartFormat
-      , eventTimeRangeEndFormat
-      , eventWrapperComponent: EventWrapper
-      , rtl: isRtl
-      , step
-      , timeslots
-      , startAccessor, endAccessor, titleAccessor } = this.props;
-
-    let EventComponent = eventComponent
+      components: { event: EventComponent },
+      culture,
+      endAccessor,
+      eventPropGetter,
+      eventTimeRangeEndFormat,
+      eventTimeRangeFormat,
+      eventTimeRangeStartFormat,
+      eventWrapperComponent: EventWrapper,
+      events,
+      max,
+      messages,
+      min,
+      rtl: isRtl,
+      selected,
+      showMultiDayTimes,
+      startAccessor,
+      step,
+      timeslots,
+      titleAccessor,
+    } = this.props
 
     let styledEvents = getStyledEvents({
-      events, startAccessor, endAccessor, min, showMultiDayTimes, totalMin: this._totalMin, step, timeslots
+      events,
+      startAccessor,
+      endAccessor,
+      min,
+      showMultiDayTimes,
+      totalMin: this._totalMin,
+      step,
+      timeslots,
     })
 
     return styledEvents.map(({ event, style }, idx) => {
-      let _eventTimeRangeFormat = eventTimeRangeFormat;
-      let _continuesPrior = false;
-      let _continuesAfter = false;
+      let _eventTimeRangeFormat = eventTimeRangeFormat
+      let _continuesPrior = false
+      let _continuesAfter = false
       let start = get(event, startAccessor)
       let end = get(event, endAccessor)
 
       if (start < min) {
-        start = min;
-        _continuesPrior = true;
-        _eventTimeRangeFormat = eventTimeRangeEndFormat;
+        start = min
+        _continuesPrior = true
+        _eventTimeRangeFormat = eventTimeRangeEndFormat
       }
 
       if (end > max) {
-        end = max;
-        _continuesAfter = true;
-        _eventTimeRangeFormat = eventTimeRangeStartFormat;
+        end = max
+        _continuesAfter = true
+        _eventTimeRangeFormat = eventTimeRangeStartFormat
       }
 
       let continuesPrior = startsBefore(start, min)
       let continuesAfter = startsAfter(end, max)
 
       let title = get(event, titleAccessor)
-      let label;
+      let label
       if (_continuesPrior && _continuesAfter) {
-        label = messages.allDay;
+        label = messages.allDay
       } else {
-          label = localizer.format({start, end}, _eventTimeRangeFormat, culture);
+        label = localizer.format({ start, end }, _eventTimeRangeFormat, culture)
       }
 
       let _isSelected = isSelected(event, selected)
 
       if (eventPropGetter)
-        var { style: xStyle, className } = eventPropGetter(event, start, end, _isSelected)
+        var { style: xStyle, className } = eventPropGetter(
+          event,
+          start,
+          end,
+          _isSelected
+        )
 
       let { height, top, width, xOffset } = style
 
@@ -205,58 +228,59 @@ class DayColumn extends React.Component {
               top: `${top}%`,
               height: `${height}%`,
               [isRtl ? 'right' : 'left']: `${Math.max(0, xOffset)}%`,
-              width: `${width}%`
+              width: `${width}%`,
             }}
-            title={(typeof label === 'string' ? label + ': ' : '') + title }
-            onClick={(e) => this._select(event, e)}
-            onDoubleClick={(e) => this._doubleClick(event, e)}
+            title={(typeof label === 'string' ? label + ': ' : '') + title}
+            onClick={e => this._select(event, e)}
+            onDoubleClick={e => this._doubleClick(event, e)}
             className={cn('rbc-event', className, {
               'rbc-selected': _isSelected,
               'rbc-event-continues-earlier': continuesPrior,
               'rbc-event-continues-later': continuesAfter,
               'rbc-event-continues-day-prior': _continuesPrior,
-              'rbc-event-continues-day-after': _continuesAfter
+              'rbc-event-continues-day-after': _continuesAfter,
             })}
           >
-            <div className='rbc-event-label'>{label}</div>
-            <div className='rbc-event-content'>
-              { EventComponent
-                ? <EventComponent event={event} title={title}/>
-                : title
-              }
+            <div className="rbc-event-label">{label}</div>
+            <div className="rbc-event-content">
+              {EventComponent ? (
+                <EventComponent event={event} title={title} />
+              ) : (
+                title
+              )}
             </div>
           </div>
         </EventWrapper>
       )
     })
-  };
+  }
 
   _slotStyle = (startSlot, endSlot) => {
-    let top = ((startSlot / this._totalMin) * 100);
-    let bottom = ((endSlot / this._totalMin) * 100);
+    let top = startSlot / this._totalMin * 100
+    let bottom = endSlot / this._totalMin * 100
 
     return {
       top: top + '%',
-      height: bottom - top + '%'
+      height: bottom - top + '%',
     }
-  };
+  }
 
   _selectable = () => {
-    let node = findDOMNode(this);
-    let selector = this._selector = new Selection(()=> findDOMNode(this), {
+    let node = findDOMNode(this)
+    let selector = (this._selector = new Selection(() => findDOMNode(this), {
       longPressThreshold: this.props.longPressThreshold,
-    })
+    }))
 
-    let maybeSelect = (box) => {
+    let maybeSelect = box => {
       let onSelecting = this.props.onSelecting
-      let current = this.state || {};
-      let state = selectionState(box);
-      let { startDate: start, endDate: end } = state;
+      let current = this.state || {}
+      let state = selectionState(box)
+      let { startDate: start, endDate: end } = state
 
       if (onSelecting) {
         if (
           (dates.eq(current.startDate, start, 'minutes') &&
-          dates.eq(current.endDate, end, 'minutes')) ||
+            dates.eq(current.endDate, end, 'minutes')) ||
           onSelecting({ start, end }) === false
         )
           return
@@ -266,21 +290,20 @@ class DayColumn extends React.Component {
     }
 
     let selectionState = ({ y }) => {
-      let { step, min, max } = this.props;
+      let { step, min, max } = this.props
       let { top, bottom } = getBoundsForNode(node)
 
-      let mins = this._totalMin;
+      let mins = this._totalMin
 
       let range = Math.abs(top - bottom)
 
-      let current = (y - top) / range;
+      let current = (y - top) / range
 
       current = snapToSlot(minToDate(mins * current, min), step)
 
-      if (!this.state.selecting)
-        this._initialDateSlot = current
+      if (!this.state.selecting) this._initialDateSlot = current
 
-      let initial = this._initialDateSlot;
+      let initial = this._initialDateSlot
 
       if (dates.eq(initial, current, 'minutes'))
         current = dates.add(current, step, 'minutes')
@@ -293,7 +316,7 @@ class DayColumn extends React.Component {
         startDate: start,
         endDate: end,
         startSlot: positionFromDate(start, min, this._totalMin),
-        endSlot: positionFromDate(end, min, this._totalMin)
+        endSlot: positionFromDate(end, min, this._totalMin),
       }
     }
 
@@ -307,36 +330,33 @@ class DayColumn extends React.Component {
     selector.on('selecting', maybeSelect)
     selector.on('selectStart', maybeSelect)
 
-    selector.on('beforeSelect', (box) => {
+    selector.on('beforeSelect', box => {
       if (this.props.selectable !== 'ignoreEvents') return
 
       return !isEvent(findDOMNode(this), box)
     })
 
-    selector
-      .on('click', box => selectorClicksHandler(box, 'click'))
+    selector.on('click', box => selectorClicksHandler(box, 'click'))
 
-    selector
-      .on('doubleClick', (box) => selectorClicksHandler(box, 'doubleClick'))
+    selector.on('doubleClick', box => selectorClicksHandler(box, 'doubleClick'))
 
-    selector
-      .on('select', () => {
-        if (this.state.selecting) {
-          this._selectSlot({ ...this.state, action: 'select' })
-          this.setState({ selecting: false })
-        }
-      })
-  };
+    selector.on('select', () => {
+      if (this.state.selecting) {
+        this._selectSlot({ ...this.state, action: 'select' })
+        this.setState({ selecting: false })
+      }
+    })
+  }
 
   _teardownSelectable = () => {
     if (!this._selector) return
-    this._selector.teardown();
-    this._selector = null;
-  };
+    this._selector.teardown()
+    this._selector = null
+  }
 
   _selectSlot = ({ startDate, endDate, action }) => {
-    let current = startDate
-      , slots = [];
+    let current = startDate,
+      slots = []
 
     while (dates.lte(current, endDate)) {
       slots.push(current)
@@ -347,28 +367,28 @@ class DayColumn extends React.Component {
       slots,
       start: startDate,
       end: endDate,
-      action
+      resourceId: this.props.resource,
+      action,
     })
-  };
+  }
 
   _select = (...args) => {
     notify(this.props.onSelectEvent, args)
-  };
+  }
 
   _doubleClick = (...args) => {
     notify(this.props.onDoubleClickEvent, args)
-  };
+  }
 }
 
+function minToDate(min, date) {
+  var dt = new Date(date),
+    totalMins = dates.diff(dates.startOf(date, 'day'), date, 'minutes')
 
-function minToDate(min, date){
-  var dt = new Date(date)
-    , totalMins = dates.diff(dates.startOf(date, 'day'), date, 'minutes');
-
-  dt = dates.hours(dt, 0);
-  dt = dates.minutes(dt, totalMins + min);
+  dt = dates.hours(dt, 0)
+  dt = dates.minutes(dt, totalMins + min)
   dt = dates.seconds(dt, 0)
   return dates.milliseconds(dt, 0)
 }
 
-export default DayColumn;
+export default DayColumn
