@@ -1,9 +1,9 @@
-import PropTypes from 'prop-types';
-import React from 'react';
-import cn from 'classnames';
-import dates from './utils/dates';
-import { accessor, elementType } from './utils/propTypes';
-import { accessor as get } from './utils/accessors';
+import PropTypes from 'prop-types'
+import React from 'react'
+import cn from 'classnames'
+import dates from './utils/dates'
+import { accessor, elementType } from './utils/propTypes'
+import { accessor as get } from './utils/accessors'
 
 let propTypes = {
   event: PropTypes.object.isRequired,
@@ -11,6 +11,7 @@ let propTypes = {
   slotEnd: PropTypes.instanceOf(Date),
 
   selected: PropTypes.bool,
+  isAllDay: PropTypes.bool,
   eventPropGetter: PropTypes.func,
   titleAccessor: accessor,
   allDayAccessor: accessor,
@@ -20,60 +21,73 @@ let propTypes = {
   eventComponent: elementType,
   eventWrapperComponent: elementType.isRequired,
   onSelect: PropTypes.func,
-  onDoubleClick: PropTypes.func
+  onDoubleClick: PropTypes.func,
 }
 
 class EventCell extends React.Component {
   render() {
     let {
-        className
-      , event
-      , selected
-      , eventPropGetter
-      , startAccessor, endAccessor, titleAccessor
-      , slotStart
-      , slotEnd
-      , onSelect
-      , onDoubleClick
-      , eventComponent: Event
-      , eventWrapperComponent: EventWrapper
-      , ...rest } = this.props;
+      className,
+      event,
+      selected,
+      isAllDay,
+      eventPropGetter,
+      startAccessor,
+      endAccessor,
+      titleAccessor,
+      slotStart,
+      slotEnd,
+      onSelect,
+      onDoubleClick,
+      eventComponent: Event,
+      eventWrapperComponent: EventWrapper,
+      ...props
+    } = this.props
 
-    let title = get(event, titleAccessor)
-      , end = get(event, endAccessor)
-      , start = get(event, startAccessor)
-      , isAllDay = get(event, props.allDayAccessor)
-      , continuesPrior = dates.lt(start, slotStart, 'day')
-      , continuesAfter = dates.gte(end, slotEnd, 'day')
+    let title = get(event, titleAccessor),
+      end = get(event, endAccessor),
+      start = get(event, startAccessor),
+      isAllDayEvent =
+        isAllDay ||
+        get(event, props.allDayAccessor) ||
+        dates.diff(start, dates.ceil(end, 'day'), 'day') > 1,
+      continuesPrior = dates.lt(start, slotStart, 'day'),
+      continuesAfter = dates.gte(end, slotEnd, 'day')
 
     if (eventPropGetter)
-      var { style, className: xClassName } = eventPropGetter(event, start, end, selected);
+      var { style, className: xClassName } = eventPropGetter(
+        event,
+        start,
+        end,
+        selected
+      )
 
     return (
       <EventWrapper event={event}>
         <div
-          style={{...rest.style, ...style}}
+          style={{ ...props.style, ...style }}
           className={cn('rbc-event', className, xClassName, {
             'rbc-selected': selected,
-            'rbc-event-allday': isAllDay || dates.diff(start, dates.ceil(end, 'day'), 'day') > 1,
+            'rbc-event-allday': isAllDayEvent,
             'rbc-event-continues-prior': continuesPrior,
-            'rbc-event-continues-after': continuesAfter
+            'rbc-event-continues-after': continuesAfter,
           })}
-          onClick={(e) => onSelect(event, e)}
-          onDoubleClick={(e) => onDoubleClick(event, e)}
+          onClick={e => onSelect(event, e)}
+          onDoubleClick={e => onDoubleClick(event, e)}
         >
-          <div className='rbc-event-content' title={title}>
-            { Event
-              ? <Event event={event} title={title}/>
-              : title
-            }
+          <div className="rbc-event-content" title={title}>
+            {Event ? (
+              <Event event={event} title={title} isAllDay={isAllDayEvent} />
+            ) : (
+              title
+            )}
           </div>
         </div>
       </EventWrapper>
-    );
+    )
   }
 }
 
-EventCell.propTypes = propTypes;
+EventCell.propTypes = propTypes
 
 export default EventCell
