@@ -1,11 +1,11 @@
-import PropTypes from 'prop-types';
-import React from 'react';
-import { findDOMNode } from 'react-dom';
-import EventCell from './EventCell';
-import getHeight from 'dom-helpers/query/height';
-import { accessor, elementType } from './utils/propTypes';
-import { segStyle } from './utils/eventLevels';
-import { isSelected } from './utils/selection';
+import PropTypes from 'prop-types'
+import React from 'react'
+import { findDOMNode } from 'react-dom'
+import EventCell from './EventCell'
+import getHeight from 'dom-helpers/query/height'
+import { accessor, elementType } from './utils/propTypes'
+import { segStyle } from './utils/eventLevels'
+import { isSelected } from './utils/selection'
 
 /* eslint-disable react/prop-types */
 export default {
@@ -15,6 +15,7 @@ export default {
     start: PropTypes.instanceOf(Date),
 
     selected: PropTypes.object,
+    isAllDay: PropTypes.bool,
     eventPropGetter: PropTypes.func,
     titleAccessor: accessor,
     allDayAccessor: accessor,
@@ -24,23 +25,31 @@ export default {
     eventComponent: elementType,
     eventWrapperComponent: elementType.isRequired,
     onSelect: PropTypes.func,
-    onDoubleClick: PropTypes.func
+    onDoubleClick: PropTypes.func,
   },
 
   defaultProps: {
     segments: [],
     selected: {},
-    slots: 7
+    slots: 7,
   },
 
   renderEvent(props, event) {
     let {
-        eventPropGetter, selected, start, end
-      , startAccessor, endAccessor, titleAccessor
-      , allDayAccessor, eventComponent
-      , eventWrapperComponent
-      , onSelect
-      , onDoubleClick } = props;
+      eventPropGetter,
+      selected,
+      isAllDay,
+      start,
+      end,
+      startAccessor,
+      endAccessor,
+      titleAccessor,
+      allDayAccessor,
+      eventComponent,
+      eventWrapperComponent,
+      onSelect,
+      onDoubleClick,
+    } = props
 
     return (
       <EventCell
@@ -50,6 +59,7 @@ export default {
         onSelect={onSelect}
         onDoubleClick={onDoubleClick}
         selected={isSelected(event, selected)}
+        isAllDay={isAllDay}
         startAccessor={startAccessor}
         endAccessor={endAccessor}
         titleAccessor={titleAccessor}
@@ -61,17 +71,21 @@ export default {
     )
   },
 
-  renderSpan(props, len, key, content = ' '){
-    let { slots } = props;
+  renderSpan(props, len, key, content = ' ') {
+    let { slots } = props
 
     return (
-      <div key={key} className='rbc-row-segment' style={segStyle(Math.abs(len), slots)}>
+      <div
+        key={key}
+        className="rbc-row-segment"
+        style={segStyle(Math.abs(len), slots)}
+      >
         {content}
       </div>
     )
   },
 
-  getRowHeight(){
+  getRowHeight() {
     getHeight(findDOMNode(this))
-  }
+  },
 }
