@@ -1,9 +1,9 @@
-import React from 'react';
+import React from 'react'
 
 class EventWrapper extends React.Component {
   render() {
-    return this.props.children;
+    return this.props.children
   }
 }
 
-export default EventWrapper;
+export default EventWrapper
