@@ -1,42 +1,41 @@
-import PropTypes from 'prop-types';
-import React, { Component } from 'react';
-import cn from 'classnames';
-import { findDOMNode } from 'react-dom';
+import PropTypes from 'prop-types'
+import React, { Component } from 'react'
+import cn from 'classnames'
+import { findDOMNode } from 'react-dom'
 
-import dates from './utils/dates';
+import dates from './utils/dates'
 import localizer from './localizer'
-import DayColumn from './DayColumn';
-import TimeColumn from './TimeColumn';
-import DateContentRow from './DateContentRow';
-import Header from './Header';
+import DayColumn from './DayColumn'
+import TimeColumn from './TimeColumn'
+import DateContentRow from './DateContentRow'
+import Header from './Header'
 
-import getWidth from 'dom-helpers/query/width';
-import scrollbarSize from 'dom-helpers/util/scrollbarSize';
-import message from './utils/messages';
+import getWidth from 'dom-helpers/query/width'
+import scrollbarSize from 'dom-helpers/util/scrollbarSize'
+import message from './utils/messages'
 
-import { accessor, dateFormat } from './utils/propTypes';
+import { accessor, dateFormat } from './utils/propTypes'
 
-import { notify } from './utils/helpers';
+import { notify } from './utils/helpers'
 
-import { accessor as get } from './utils/accessors';
+import { accessor as get } from './utils/accessors'
 
-import { inRange, sortEvents, segStyle } from './utils/eventLevels';
+import { inRange, sortEvents, segStyle } from './utils/eventLevels'
 
 export default class TimeGrid extends Component {
-
   static propTypes = {
     events: PropTypes.array.isRequired,
+    resources: PropTypes.array,
 
     step: PropTypes.number,
-    range: PropTypes.arrayOf(
-      PropTypes.instanceOf(Date)
-    ),
+    range: PropTypes.arrayOf(PropTypes.instanceOf(Date)),
     min: PropTypes.instanceOf(Date),
     max: PropTypes.instanceOf(Date),
     now: PropTypes.instanceOf(Date),
 
     scrollToTime: PropTypes.instanceOf(Date),
     eventPropGetter: PropTypes.func,
+    dayPropGetter: PropTypes.func,
     dayFormat: dateFormat,
     showMultiDayTimes: PropTypes.bool,
     culture: PropTypes.string,
@@ -48,6 +47,10 @@ export default class TimeGrid extends Component {
     allDayAccessor: accessor.isRequired,
     startAccessor: accessor.isRequired,
     endAccessor: accessor.isRequired,
+    resourceAccessor: accessor.isRequired,
+
+    resourceIdAccessor: accessor.isRequired,
+    resourceTitleAccessor: accessor.isRequired,
 
     selected: PropTypes.object,
     selectable: PropTypes.oneOf([true, false, 'ignoreEvents']),
@@ -75,36 +78,36 @@ export default class TimeGrid extends Component {
      * There is a strange bug in React, using ...TimeColumn.defaultProps causes weird crashes
      */
     type: 'gutter',
-    now: new Date()
+    now: new Date(),
   }
 
   constructor(props) {
     super(props)
-    this.state = { gutterWidth: undefined, isOverflowing: null };
+    this.state = { gutterWidth: undefined, isOverflowing: null }
     this.handleSelectEvent = this.handleSelectEvent.bind(this)
     this.handleDoubleClickEvent = this.handleDoubleClickEvent.bind(this)
     this.handleHeaderClick = this.handleHeaderClick.bind(this)
   }
 
   componentWillMount() {
-    this._gutters = [];
-    this.calculateScroll();
+    this._gutters = []
+    this.calculateScroll()
   }
 
   componentDidMount() {
-    this.checkOverflow();
+    this.checkOverflow()
 
     if (this.props.width == null) {
       this.measureGutter()
     }
-    this.applyScroll();
+    this.applyScroll()
 
-    this.positionTimeIndicator();
-    this.triggerTimeIndicatorUpdate();
+    this.positionTimeIndicator()
+    this.triggerTimeIndicatorUpdate()
   }
 
   componentWillUnmount() {
-    window.clearTimeout(this._timeIndicatorTimeout);
+    window.clearTimeout(this._timeIndicatorTimeout)
   }
 
   componentDidUpdate() {
@@ -112,60 +115,64 @@ export default class TimeGrid extends Component {
       this.measureGutter()
     }
 
-    this.applyScroll();
-    this.positionTimeIndicator();
+    this.applyScroll()
+    this.positionTimeIndicator()
     //this.checkOverflow()
   }
 
   componentWillReceiveProps(nextProps) {
-    const { range, scrollToTime } = this.props;
+    const { range, scrollToTime } = this.props
     // When paginating, reset scroll
     if (
       !dates.eq(nextProps.range[0], range[0], 'minute') ||
       !dates.eq(nextProps.scrollToTime, scrollToTime, 'minute')
     ) {
-      this.calculateScroll();
+      this.calculateScroll()
     }
   }
 
   handleSelectAllDaySlot = (slots, slotInfo) => {
-    const { onSelectSlot } = this.props;
+    const { onSelectSlot } = this.props
     notify(onSelectSlot, {
       slots,
       start: slots[0],
       end: slots[slots.length - 1],
-      action: slotInfo.action
+      action: slotInfo.action,
     })
   }
 
   render() {
     let {
-        events
-      , range
-      , width
-      , startAccessor
-      , endAccessor
-      , allDayAccessor
-      , showMultiDayTimes} = this.props;
+      events,
+      range,
+      width,
+      startAccessor,
+      endAccessor,
+      resources,
+      allDayAccessor,
+      showMultiDayTimes,
+    } = this.props
 
-    width = width || this.state.gutterWidth;
+    width = width || this.state.gutterWidth
 
-    let start = range[0]
-      , end = range[range.length - 1]
+    let start = range[0],
+      end = range[range.length - 1]
 
-    this.slots = range.length;
+    this.slots = range.length
 
-    let allDayEvents = []
-      , rangeEvents = [];
+    let allDayEvents = [],
+      rangeEvents = []
 
     events.forEach(event => {
       if (inRange(event, start, end, this.props)) {
-        let eStart = get(event, startAccessor)
-          , eEnd = get(event, endAccessor);
+        let eStart = get(event, startAccessor),
+          eEnd = get(event, endAccessor)
 
-        if (get(event, allDayAccessor)
-          || (dates.isJustDate(eStart) && dates.isJustDate(eEnd))
-          || (!showMultiDayTimes && !dates.eq(eStart, eEnd, 'day'))) {
+        if (
+          get(event, allDayAccessor) ||
+          (dates.isJustDate(eStart) && dates.isJustDate(eEnd)) ||
+          (!showMultiDayTimes && !dates.eq(eStart, eEnd, 'day'))
+        ) {
           allDayEvents.push(event)
         } else {
           rangeEvents.push(event)
@@ -175,90 +182,119 @@ export default class TimeGrid extends Component {
 
     allDayEvents.sort((a, b) => sortEvents(a, b, this.props))
 
-    let gutterRef = ref => this._gutters[1] = ref && findDOMNode(ref);
+    let gutterRef = ref => (this._gutters[1] = ref && findDOMNode(ref))
+
+    let eventsRendered = this.renderEvents(
+      range,
+      rangeEvents,
+      this.props.now,
+      resources || [null]
+    )
 
     return (
-      <div className='rbc-time-view'>
+      <div className="rbc-time-view">
+        {this.renderHeader(range, allDayEvents, width, resources)}
 
-        {this.renderHeader(range, allDayEvents, width)}
-
-        <div ref='content' className='rbc-time-content'>
-          <div ref='timeIndicator' className='rbc-current-time-indicator' />
+        <div ref="content" className="rbc-time-content">
+          <div ref="timeIndicator" className="rbc-current-time-indicator" />
 
           <TimeColumn
             {...this.props}
             showLabels
             style={{ width }}
             ref={gutterRef}
-            className='rbc-time-gutter'
+            className="rbc-time-gutter"
           />
-
-          {this.renderEvents(range, rangeEvents, this.props.now)}
-
+          {eventsRendered}
         </div>
       </div>
-    );
+    )
   }
-
-  renderEvents(range, events, today){
-    let { min, max, endAccessor, startAccessor, components } = this.props;
+  renderEvents(range, events, today, resources) {
+    let {
+      min,
+      max,
+      endAccessor,
+      startAccessor,
+      resourceAccessor,
+      resourceIdAccessor,
+      components,
+    } = this.props
 
     return range.map((date, idx) => {
-      let daysEvents = events.filter(
-        event => dates.inRange(date,
+      let daysEvents = events.filter(event =>
+        dates.inRange(
+          date,
           get(event, startAccessor),
-          get(event, endAccessor), 'day')
+          get(event, endAccessor),
+          'day'
+        )
       )
 
-      return (
-        <DayColumn
-          {...this.props }
-          min={dates.merge(date, min)}
-          max={dates.merge(date, max)}
-          eventComponent={components.event}
-          eventWrapperComponent={components.eventWrapper}
-          dayWrapperComponent={components.dayWrapper}
-          className={cn({ 'rbc-now': dates.eq(date, today, 'day') })}
-          style={segStyle(1, this.slots)}
-          key={idx}
-          date={date}
-          events={daysEvents}
-        />
-      )
+      return resources.map((resource, id) => {
+        let eventsToDisplay = !resource
+          ? daysEvents
+          : daysEvents.filter(
+              event =>
+                get(event, resourceAccessor) ===
+                get(resource, resourceIdAccessor)
+            )
+
+        return (
+          <DayColumn
+            {...this.props}
+            min={dates.merge(date, min)}
+            max={dates.merge(date, max)}
+            resource={resource && resource.id}
+            eventComponent={components.event}
+            eventWrapperComponent={components.eventWrapper}
+            dayWrapperComponent={components.dayWrapper}
+            className={cn({ 'rbc-now': dates.eq(date, today, 'day') })}
+            style={segStyle(1, this.slots)}
+            key={idx + '-' + id}
+            date={date}
+            events={eventsToDisplay}
+          />
+        )
+      })
     })
   }
 
-  renderHeader(range, events, width) {
-    let { messages, rtl, selectable, components, now } = this.props;
-    let { isOverflowing } = this.state || {};
+  renderHeader(range, events, width, resources) {
+    let { messages, rtl, selectable, components, now } = this.props
+    let { isOverflowing } = this.state || {}
 
-    let style = {};
+    let style = {}
     if (isOverflowing)
-      style[rtl ? 'marginLeft' : 'marginRight'] = scrollbarSize() + 'px';
+      style[rtl ? 'marginLeft' : 'marginRight'] = scrollbarSize() + 'px'
+
+    let headerRendered = resources
+      ? this.renderHeaderResources(range, resources)
+      : message(messages).allDay
 
     return (
       <div
-        ref='headerCell'
-        className={cn(
-          'rbc-time-header',
-          isOverflowing && 'rbc-overflowing'
-        )}
+        ref="headerCell"
+        className={cn('rbc-time-header', isOverflowing && 'rbc-overflowing')}
         style={style}
       >
-        <div className='rbc-row'>
-          <div
-            className='rbc-label rbc-header-gutter'
-            style={{ width }}
-          />
-          { this.renderHeaderCells(range) }
+        <div className="rbc-row">
+          <div className="rbc-label rbc-header-gutter" style={{ width }} />
+          {this.renderHeaderCells(range)}
         </div>
-        <div className='rbc-row'>
+        {resources && (
+          <div className="rbc-row rbc-row-resource">
+            <div className="rbc-label rbc-header-gutter" style={{ width }} />
+            {headerRendered}
+          </div>
+        )}
+        <div className="rbc-row">
           <div
-            ref={ref => this._gutters[0] = ref}
-            className='rbc-label rbc-header-gutter'
+            ref={ref => (this._gutters[0] = ref)}
+            className="rbc-label rbc-header-gutter"
             style={{ width }}
           >
-            { message(messages).allDay }
+            {message(messages).allDay}
           </div>
           <DateContentRow
             now={now}
@@ -266,10 +302,11 @@ export default class TimeGrid extends Component {
             range={range}
             rtl={this.props.rtl}
             events={events}
-            className='rbc-allday-cell'
+            className="rbc-allday-cell"
             selectable={selectable}
             onSelectSlot={this.handleSelectAllDaySlot}
             dateCellWrapper={components.dateCellWrapper}
+            dayPropGetter={this.props.dayPropGetter}
             eventComponent={this.props.components.event}
             eventWrapperComponent={this.props.components.eventWrapper}
             titleAccessor={this.props.titleAccessor}
@@ -278,6 +315,7 @@ export default class TimeGrid extends Component {
             allDayAccessor={this.props.allDayAccessor}
             eventPropGetter={this.props.eventPropGetter}
             selected={this.props.selected}
+            isAllDay={true}
             onSelect={this.handleSelectEvent}
             onDoubleClick={this.handleDoubleClickEvent}
             longPressThreshold={this.props.longPressThreshold}
@@ -287,13 +325,39 @@ export default class TimeGrid extends Component {
     )
   }
 
-  renderHeaderCells(range){
-    let { dayFormat, culture, components, getDrilldownView } = this.props;
+  renderHeaderResources(range, resources) {
+    const { resourceTitleAccessor } = this.props
+    return range.map((date, i) => {
+      return resources.map((resource, j) => {
+        return (
+          <div
+            key={i + '-' + j}
+            className={cn('rbc-header', dates.isToday(date) && 'rbc-today')}
+            style={segStyle(1, this.slots)}
+          >
+            <span>{get(resource, resourceTitleAccessor)}</span>
+          </div>
+        )
+      })
+    })
+  }
+
+  renderHeaderCells(range) {
+    let {
+      dayFormat,
+      culture,
+      components,
+      dayPropGetter,
+      getDrilldownView,
+    } = this.props
     let HeaderComponent = components.header || Header
 
     return range.map((date, i) => {
-      let drilldownView = getDrilldownView(date);
-      let label = localizer.format(date, dayFormat, culture);
+      let drilldownView = getDrilldownView(date)
+      let label = localizer.format(date, dayFormat, culture)
+
+      const { className, style: dayStyles } =
+        (dayPropGetter && dayPropGetter(date)) || {}
 
       let header = (
         <HeaderComponent
@@ -310,28 +374,27 @@ export default class TimeGrid extends Component {
           key={i}
           className={cn(
             'rbc-header',
-            dates.isToday(date) && 'rbc-today',
+            className,
+            dates.isToday(date) && 'rbc-today'
           )}
-          style={segStyle(1, this.slots)}
+          style={Object.assign({}, dayStyles, segStyle(1, this.slots))}
         >
           {drilldownView ? (
             <a
-              href='#'
+              href="#"
               onClick={e => this.handleHeaderClick(date, drilldownView, e)}
             >
               {header}
             </a>
           ) : (
-            <span>
-              {header}
-            </span>
+            <span>{header}</span>
           )}
         </div>
       )
     })
   }
 
-  handleHeaderClick(date, view, e){
+  handleHeaderClick(date, view, e) {
     e.preventDefault()
     notify(this.props.onDrillDown, [date, view])
   }
@@ -350,17 +413,17 @@ export default class TimeGrid extends Component {
     notify(this.props.onSelectEvent, args)
   }
 
-  clearSelection(){
+  clearSelection() {
     clearTimeout(this._selectTimer)
-    this._pendingSelection = [];
+    this._pendingSelection = []
   }
 
   measureGutter() {
-    let width = this.state.gutterWidth;
-    let gutterCells = this._gutters;
+    let width = this.state.gutterWidth
+    let gutterCells = this._gutters
 
     if (!width) {
-      width = Math.max(...gutterCells.map(getWidth));
+      width = Math.max(...gutterCells.map(getWidth))
 
       if (width) {
         this.setState({ gutterWidth: width })
@@ -370,65 +433,67 @@ export default class TimeGrid extends Component {
 
   applyScroll() {
     if (this._scrollRatio) {
-      const { content } = this.refs;
-      content.scrollTop = content.scrollHeight * this._scrollRatio;
+      const { content } = this.refs
+      content.scrollTop = content.scrollHeight * this._scrollRatio
       // Only do this once
-      this._scrollRatio = null;
+      this._scrollRatio = null
     }
   }
 
   calculateScroll() {
-    const { min, max, scrollToTime } = this.props;
+    const { min, max, scrollToTime } = this.props
 
-    const diffMillis = scrollToTime - dates.startOf(scrollToTime, 'day');
-    const totalMillis = dates.diff(max, min);
+    const diffMillis = scrollToTime - dates.startOf(scrollToTime, 'day')
+    const totalMillis = dates.diff(max, min)
 
-    this._scrollRatio = diffMillis / totalMillis;
+    this._scrollRatio = diffMillis / totalMillis
   }
 
   checkOverflow() {
-    if (this._updatingOverflow) return;
+    if (this._updatingOverflow) return
 
-    let isOverflowing = this.refs.content.scrollHeight > this.refs.content.clientHeight;
+    let isOverflowing =
+      this.refs.content.scrollHeight > this.refs.content.clientHeight
 
     if (this.state.isOverflowing !== isOverflowing) {
-      this._updatingOverflow = true;
+      this._updatingOverflow = true
       this.setState({ isOverflowing }, () => {
-        this._updatingOverflow = false;
+        this._updatingOverflow = false
       })
     }
   }
 
   positionTimeIndicator() {
     const { rtl, min, max } = this.props
-    const now = new Date();
+    const now = new Date()
 
-    const secondsGrid = dates.diff(max, min, 'seconds');
-    const secondsPassed = dates.diff(now, min, 'seconds');
+    const secondsGrid = dates.diff(max, min, 'seconds')
+    const secondsPassed = dates.diff(now, min, 'seconds')
 
-    const timeIndicator = this.refs.timeIndicator;
-    const factor = secondsPassed / secondsGrid;
-    const timeGutter = this._gutters[this._gutters.length - 1];
+    const timeIndicator = this.refs.timeIndicator
+    const factor = secondsPassed / secondsGrid
+    const timeGutter = this._gutters[this._gutters.length - 1]
 
     if (timeGutter && now >= min && now <= max) {
-      const pixelHeight = timeGutter.offsetHeight;
-      const offset = Math.floor(factor * pixelHeight);
+      const pixelHeight = timeGutter.offsetHeight
+      const offset = Math.floor(factor * pixelHeight)
 
-      timeIndicator.style.display = 'block';
-      timeIndicator.style[rtl ? 'left' : 'right'] = 0;
-      timeIndicator.style[rtl ? 'right' : 'left'] = timeGutter.offsetWidth + 'px';
-      timeIndicator.style.top = offset + 'px';
+      timeIndicator.style.display = 'block'
+      timeIndicator.style[rtl ? 'left' : 'right'] = 0
+      timeIndicator.style[rtl ? 'right' : 'left'] =
+        timeGutter.offsetWidth + 'px'
+      timeIndicator.style.top = offset + 'px'
     } else {
-      timeIndicator.style.display = 'none';
+      timeIndicator.style.display = 'none'
     }
   }
 
   triggerTimeIndicatorUpdate() {
     // Update the position of the time indicator every minute
     this._timeIndicatorTimeout = window.setTimeout(() => {
-      this.positionTimeIndicator();
+      this.positionTimeIndicator()
 
-      this.triggerTimeIndicatorUpdate();
+      this.triggerTimeIndicatorUpdate()
     }, 60000)
   }
 }
