@@ -1,14 +1,12 @@
-import PropTypes from 'prop-types';
-import React from 'react';
-import cn from 'classnames';
-import { navigate } from './utils/constants';
+import PropTypes from 'prop-types'
+import React from 'react'
+import cn from 'classnames'
+import { navigate } from './utils/constants'
 
 class Toolbar extends React.Component {
   static propTypes = {
     view: PropTypes.string.isRequired,
-    views: PropTypes.arrayOf(
-      PropTypes.string,
-    ).isRequired,
+    views: PropTypes.arrayOf(PropTypes.string).isRequired,
     label: PropTypes.node.isRequired,
     messages: PropTypes.object,
     onNavigate: PropTypes.func.isRequired,
@@ -16,49 +14,43 @@ class Toolbar extends React.Component {
   }
 
   render() {
-    let { messages, label } = this.props;
+    let { messages, label } = this.props
 
     return (
-      <div className='rbc-toolbar'>
-        <span className='rbc-btn-group'>
+      <div className="rbc-toolbar">
+        <span className="rbc-btn-group">
           <button
-            type='button'
+            type="button"
             onClick={this.navigate.bind(null, navigate.TODAY)}
           >
             {messages.today}
           </button>
           <button
-            type='button'
+            type="button"
             onClick={this.navigate.bind(null, navigate.PREVIOUS)}
           >
             {messages.previous}
           </button>
           <button
-            type='button'
+            type="button"
             onClick={this.navigate.bind(null, navigate.NEXT)}
           >
             {messages.next}
           </button>
         </span>
 
-        <span className='rbc-toolbar-label'>
-          { label }
-        </span>
+        <span className="rbc-toolbar-label">{label}</span>
 
-        <span className='rbc-btn-group'>
-        {
-          this.viewNamesGroup(messages)
-        }
-        </span>
+        <span className="rbc-btn-group">{this.viewNamesGroup(messages)}</span>
       </div>
-    );
+    )
   }
 
-  navigate = (action) => {
+  navigate = action => {
     this.props.onNavigate(action)
   }
 
-  view = (view) => {
+  view = view => {
     this.props.onViewChange(view)
   }
 
@@ -67,18 +59,18 @@ class Toolbar extends React.Component {
     const view = this.props.view
 
     if (viewNames.length > 1) {
-      return (
-        viewNames.map(name =>
-          <button type='button' key={name}
-            className={cn({'rbc-active': view === name})}
-            onClick={this.view.bind(null, name)}
-          >
-            {messages[name]}
-          </button>
-        )
-      )
+      return viewNames.map(name => (
+        <button
+          type="button"
+          key={name}
+          className={cn({ 'rbc-active': view === name })}
+          onClick={this.view.bind(null, name)}
+        >
+          {messages[name]}
+        </button>
+      ))
     }
   }
 }
 
-export default Toolbar;
+export default Toolbar
