@@ -1,16 +1,16 @@
-import PropTypes from 'prop-types';
-import React from 'react';
-import dates from './utils/dates';
-import localizer from './localizer';
-import { navigate } from './utils/constants';
-import TimeGrid from './TimeGrid';
+import PropTypes from 'prop-types'
+import React from 'react'
+import dates from './utils/dates'
+import localizer from './localizer'
+import { navigate } from './utils/constants'
+import TimeGrid from './TimeGrid'
 
 class Week extends React.Component {
   static propTypes = {
     date: PropTypes.instanceOf(Date).isRequired,
-  };
+  }
 
-  static defaultProps = TimeGrid.defaultProps;
+  static defaultProps = TimeGrid.defaultProps
 
   render() {
     let { date, ...rest } = this.props
@@ -23,15 +23,15 @@ class Week extends React.Component {
 }
 
 Week.navigate = (date, action) => {
-  switch (action){
+  switch (action) {
     case navigate.PREVIOUS:
-      return dates.add(date, -1, 'week');
+      return dates.add(date, -1, 'week')
 
     case navigate.NEXT:
       return dates.add(date, 1, 'week')
 
     default:
-      return date;
+      return date
   }
 }
 
@@ -43,14 +43,13 @@ Week.range = (date, { culture }) => {
   return dates.range(start, end)
 }
 
-
 Week.title = (date, { formats, culture }) => {
   let [start, ...rest] = Week.range(date, { culture })
   return localizer.format(
     { start, end: rest.pop() },
     formats.dayRangeHeaderFormat,
     culture
-  );
+  )
 }
 
 export default Week
