@@ -1,18 +1,18 @@
-import PropTypes from 'prop-types';
-import invariant from 'invariant';
+import PropTypes from 'prop-types'
+import invariant from 'invariant'
 
-const localePropType = PropTypes.oneOfType([
-        PropTypes.string,
-        PropTypes.func
-      ]);
+const localePropType = PropTypes.oneOfType([PropTypes.string, PropTypes.func])
 
 function _format(localizer, formatter, value, format, culture) {
-  let result = typeof format === 'function'
-    ? format(value, culture, localizer)
-    : formatter.call(localizer, value, format, culture)
+  let result =
+    typeof format === 'function'
+      ? format(value, culture, localizer)
+      : formatter.call(localizer, value, format, culture)
 
-  invariant(result == null || typeof result === 'string'
-    , '`localizer format(..)` must return a string, null, or undefined')
+  invariant(
+    result == null || typeof result === 'string',
+    '`localizer format(..)` must return a string, null, or undefined'
+  )
 
   return result
 }
@@ -29,15 +29,19 @@ const newFormat = (localizer, formatter) => (value, format, culture) => {
 }
 
 class DateLocalizer {
-
   constructor(spec) {
-    this.spec = spec;
-    invariant(typeof spec.format === 'function'
-      , 'date localizer `format(..)` must be a function')
-    invariant(typeof spec.parse === 'function'
-      , 'date localizer `parse(..)` must be a function')
-    invariant(typeof spec.firstOfWeek === 'function'
-      , 'date localizer `firstOfWeek(..)` must be a function')
+    invariant(
+      typeof spec.format === 'function',
+      'date localizer `format(..)` must be a function'
+    )
+    invariant(
+      typeof spec.parse === 'function',
+      'date localizer `parse(..)` must be a function'
+    )
+    invariant(
+      typeof spec.firstOfWeek === 'function',
+      'date localizer `firstOfWeek(..)` must be a function'
+    )
 
     this.propType = spec.propType || localePropType
 
@@ -51,9 +55,10 @@ class DateLocalizer {
     /*this.parse = (value, format, culture) => {
       let result = spec.parse.call(this, value, format, culture)
 
-      invariant(result == null
-        || (result instanceof Date && !isNaN(result.getTime()))
-        , 'date localizer `parse(..)` must return a valid Date, null, or undefined')
+      invariant(
+        result == null || (result instanceof Date && !isNaN(result.getTime())),
+        'date localizer `parse(..)` must return a valid Date, null, or undefined'
+      )
 
       return result
     }*/
@@ -71,38 +76,39 @@ class DateLocalizer {
 
 let localizer = {
   parse: error,
-  format:error,
-  startOfWeek: error
+  format: error,
+  startOfWeek: error,
 }
 
 export function set(newLocalizer) {
   if (!newLocalizer.__isLocalizer__) {
-    newLocalizer = new DateLocalizer(newLocalizer);
+    newLocalizer = new DateLocalizer(newLocalizer)
     newLocalizer.__isLocalizer__ = true
   }
 
-  localizer = newLocalizer;
-  return localizer;
+  localizer = newLocalizer
+  return localizer
 }
 
 let exp = {
-  parse(...args){
+  parse(...args) {
     return localizer.parse(...args)
   },
 
-  format(...args){
+  format(...args) {
     return localizer.format(...args)
   },
 
-  startOfWeek(...args){
+  startOfWeek(...args) {
     return localizer.startOfWeek(...args)
-  }
+  },
 }
 
 export default exp
 
-function error(){
+function error() {
   throw new Error(
     'You have not selected a localization strategy for Big Calendar. ' +
-    'Please use either of the two included.')
+      'Please use either of the two included.'
+  )
 }
