@@ -1,23 +1,23 @@
-import dates from '../utils/dates';
-import { set } from '../formats';
-import { set as setLocalizer } from '../localizer';
+import dates from '../utils/dates'
+import { set } from '../formats'
+import { set as setLocalizer } from '../localizer'
 
-let dateRangeFormat = ({ start, end }, culture, local)=>
+let dateRangeFormat = ({ start, end }, culture, local) =>
   local.format(start, 'd', culture) + ' — ' + local.format(end, 'd', culture)
 
-let timeRangeFormat = ({ start, end }, culture, local)=>
-  local.format(start, 't', culture) +
-    ' — ' + local.format(end, 't', culture)
+let timeRangeFormat = ({ start, end }, culture, local) =>
+  local.format(start, 't', culture) + ' — ' + local.format(end, 't', culture)
 
-let timeRangeStartFormat = ({ start, end }, culture, local)=>
+let timeRangeStartFormat = ({ start, end }, culture, local) =>
   local.format(start, 't', culture) + ' — '
 
-let timeRangeEndFormat = ({ start, end }, culture, local)=>
+let timeRangeEndFormat = ({ start, end }, culture, local) =>
   ' — ' + local.format(end, 't', culture)
 
-let weekRangeFormat = ({ start, end }, culture, local)=>
+let weekRangeFormat = ({ start, end }, culture, local) =>
   local.format(start, 'MMM dd', culture) +
-    ' - ' + local.format(end, dates.eq(start, end, 'month') ? 'dd' : 'MMM dd', culture)
+  ' - ' +
+  local.format(end, dates.eq(start, end, 'month') ? 'dd' : 'MMM dd', culture)
 
 export let formats = {
   dateFormat: 'dd',
@@ -38,15 +38,12 @@ export let formats = {
 
   agendaDateFormat: 'ddd MMM dd',
   agendaTimeFormat: 't',
-  agendaTimeRangeFormat: timeRangeFormat
+  agendaTimeRangeFormat: timeRangeFormat,
 }
 
 export default function(globalize) {
-
-  function getCulture(culture){
-    return culture
-      ? globalize.findClosestCulture(culture)
-      : globalize.culture()
+  function getCulture(culture) {
+    return culture ? globalize.findClosestCulture(culture) : globalize.culture()
   }
 
   function firstOfWeek(culture) {
@@ -59,12 +56,12 @@ export default function(globalize) {
   return setLocalizer({
     firstOfWeek,
 
-    parse(value, format, culture){
+    parse(value, format, culture) {
       return globalize.parseDate(value, format, culture)
     },
 
-    format(value, format, culture){
+    format(value, format, culture) {
       return globalize.format(value, format, culture)
-    }
+    },
   })
 }
