@@ -1,16 +1,14 @@
-
 export let navigate = {
-  PREVIOUS:  'PREV',
+  PREVIOUS: 'PREV',
   NEXT: 'NEXT',
   TODAY: 'TODAY',
-  DATE: 'DATE'
+  DATE: 'DATE',
 }
-
 
 export let views = {
   MONTH: 'month',
   WEEK: 'week',
   WORK_WEEK: 'work_week',
   DAY: 'day',
-  AGENDA: 'agenda'
+  AGENDA: 'agenda',
 }
