@@ -1,6 +1,6 @@
-import invariant from 'invariant';
+import invariant from 'invariant'
 
- let defaultMessages = {
+let defaultMessages = {
   date: 'Date',
   time: 'Time',
   event: 'Event',
@@ -16,25 +16,26 @@ import invariant from 'invariant';
   today: 'today',
   agenda: 'agenda',
 
-  showMore: total => `+${total} more`
+  showMore: total => `+${total} more`,
 }
 
-export function set(key, msg){
-  invariant(messages.hasOwnProperty(key),
+export function set(key, msg) {
+  invariant(
+    messages.hasOwnProperty(key),
     `The message key: "${key}" is not a valid message name. ` +
-    `valid keys are: ${Object.keys(messages).join(', ')}`
+      `valid keys are: ${Object.keys(messages).join(', ')}`
   )
 
-  messages[key] = msg;
+  messages[key] = msg
 }
 
-export function result(msg, ...args){
+export function result(msg, ...args) {
   return typeof msg === 'function' ? msg(args) : msg
 }
 
 export default function messages(msgs) {
   return {
     ...defaultMessages,
-    ...msgs
+    ...msgs,
   }
 }
