@@ -1,19 +1,21 @@
-import invariant from 'invariant';
-import { navigate } from './constants';
-import VIEWS from '../Views';
+import invariant from 'invariant'
+import { navigate } from './constants'
+import VIEWS from '../Views'
 
 export default function moveDate(View, { action, date, ...props }) {
-  View = typeof View === 'string' ? VIEWS[View] : View;
+  View = typeof View === 'string' ? VIEWS[View] : View
 
   switch (action) {
     case navigate.TODAY:
       date = new Date()
-      break;
+      break
     case navigate.DATE:
-      break;
+      break
     default:
-      invariant(View && typeof View.navigate === 'function',
-        'Calendar View components must implement a static `.navigate(date, action)` method.s')
+      invariant(
+        View && typeof View.navigate === 'function',
+        'Calendar View components must implement a static `.navigate(date, action)` method.s'
+      )
       date = View.navigate(date, action, props)
   }
   return date
