@@ -1,10 +1,10 @@
-import PropTypes from 'prop-types';
-import localizer from '../localizer';
-import elementType from 'react-prop-types/lib/elementType';
-import all from 'react-prop-types/lib/all';
-import { views as Views } from './constants';
+import PropTypes from 'prop-types'
+import localizer from '../localizer'
+import elementType from 'react-prop-types/lib/elementType'
+import all from 'react-prop-types/lib/all'
+import { views as Views } from './constants'
 
-import createChainableTypeChecker from 'react-prop-types/lib/utils/createChainableTypeChecker';
+import createChainableTypeChecker from 'react-prop-types/lib/utils/createChainableTypeChecker'
 
 export { elementType }
 
@@ -25,20 +25,17 @@ export let eventComponent = PropTypes.oneOfType([
     month: elementType,
     week: elementType,
     day: elementType,
-    agenda: elementType
-  })
+    agenda: elementType,
+  }),
 ])
-
 
 let viewNames = Object.keys(Views).map(k => Views[k])
 
-export let accessor = PropTypes.oneOfType([
-  PropTypes.string,
-  PropTypes.func
-])
+export let accessor = PropTypes.oneOfType([PropTypes.string, PropTypes.func])
 
 export let dateFormat = createChainableTypeChecker(
-    (...args) => localizer.propType && localizer.propType(...args))
+  (...args) => localizer.propType && localizer.propType(...args)
+)
 
 export let dateRangeFormat = PropTypes.func
 
@@ -60,24 +57,18 @@ export let dateRangeFormat = PropTypes.func
  * ```
  */
 export let views = PropTypes.oneOfType([
-  PropTypes.arrayOf(
-    PropTypes.oneOf(viewNames)
-  ),
-  all(
-    PropTypes.object,
-    (props, name, ...args)=>{
-      let prop = props[name]
-        , err;
+  PropTypes.arrayOf(PropTypes.oneOf(viewNames)),
+  all(PropTypes.object, (props, name, ...args) => {
+    let prop = props[name],
+      err
 
-      Object.keys(prop).every(key => {
-        let isBuiltinView =
-          viewNames.indexOf(key) !== -1 &&
-          typeof prop[key] === 'boolean';
+    Object.keys(prop).every(key => {
+      let isBuiltinView =
+        viewNames.indexOf(key) !== -1 && typeof prop[key] === 'boolean'
 
-        return isBuiltinView || !(err = elementType(prop, key, ...args))
-      })
+      return isBuiltinView || !(err = elementType(prop, key, ...args))
+    })
 
-      return err || null
-    }
-  )
+    return err || null
+  }),
 ])
